+++ src/index.ts
@@ -1,4 +1,5 @@
 import SimpleCalendar from "./classes/simple-calendar";
+import Macros from "./classes/macros";
 import {Logger} from "./classes/logging";
 SimpleCalendar.instance = new SimpleCalendar();
 Hooks.on('ready', () => {
@@ -6,7 +7,11 @@ Hooks.on('ready', () => {
     SimpleCalendar.instance.init().catch(Logger.error);
 
     //Expose the macro show function
-    (window as any).SimpleCalendar = {show: SimpleCalendar.instance.macroShow.bind(SimpleCalendar.instance)};
+    (window as any).SimpleCalendar = {
+        show: Macros.show,
+        setDateTime: Macros.setDateTime,
+        changeDateTime: Macros.changeDateTime
+    };
 });
 Hooks.on('getSceneControlButtons', SimpleCalendar.instance.getSceneControlButtons);
 Hooks.on("updateWorldTime", SimpleCalendar.instance.worldTimeUpdate.bind(SimpleCalendar.instance));